OSW-12: feat: move login and registration api to auth path | docs: add swagger to login api
