+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './enum';
 export * from './getParams';
 export * from './payOSUtils';
 export * from './sortObjects';
+export * from './route';
