+++ src/utils/route.ts
@@ -1,0 +1,8 @@
+export const publicRoutes = [
+  "/"
+]
+export const authRoutes = [
+  "/sign-in",
+  "/sign-up"
+]
+export const publicApiRoutesPrefix = "/api/public"
